Split image workspace orchestration hooks
